`quantum_espresso`: Fix stress units (#282)
Currently, the `extract_from_trajectory` calculation function simply
passes the stress from the last frame of the trajectory to the
`stress` output. However, the units of this output array are in GPa, not
the agreed upon eV/angstrom^3.

Here we use the `pint` library to convert the `stress` array into the
correct units.