DevOps: adopt PEP 621 and move build spec to `pyproject.toml` (#296)
Following PEP 621 it is now possible to fully define the build procedure
of your package in `pyproject.toml`. Since this PEP is now well
supported, and for example `pip` can use it, we migrate to it since it
provides a bunch of benefits:

 * No longer need the deprecated `setup.py` to install the package,
   nor the custom `setup.json`.
 * Version number and description are now dynamically fetched from the
   package. So there is no need to manually update the `setup.json`.
 * The `MANIFEST.in` is no longer necessary. The `flit` build tool will
   automatically include anything. The `pyproject.toml` has a tool
   section for `flit` that excludes the `tests` and `docs` folders.

Also the `LICENCE` file is renamed to `LICENSE.txt` to correct the typo
and use a correct file extension.