QE: Add `verification-pbe-v1` protocol
Add the `verification-pbe-v1` protocol for Quantum ESPRESSO by adding a
new `protocol.yml` file to the QE `relax` subpackage. This contains the new
protocol as a set of overrides for the default protocol used in the Quantum
ESPRESSO plugin package (`moderate`).

As the `PwRelaxWorkChain.get_builder_from_protocol()` method only accepts the
protocols defined in the plugin package, we check if the user specified protocol
is in the defaults (`fast`, `moderate`, `precise`). If not, we try to load it
from the local protocols encoded in the `aiida-common-workflows` package, and
pass the local protocol inputs as overrides to the `get_builder_from_protocol()`
method.