Add the `get_ts_energy` function for Quantum ESPRESSO
The `get_ts_energy` function extracts the "smearing contribution" to the free
energy from the `output_parameters` of the  `PwRelaxWorkChain`, where the
quantity is reported as `energy_smearing`.

Since the smearing contribution is reported as -T*S in Quantum ESPRESSO, we flip
the sign to be consistent with the`get_ts_energy` of other engines.