Update VASP part to comply with ACWF study (#303)
We add dedicated handler overrides and a dedicated protocol that was used for this study.
Also, a dedicated potential mapping has been created as the VASP developers was in the
process of updating a few potentials, in particular for the lanthanides.
Furthermore, we disable the present default to hard stop on CNORMN warnings for this study.
This goes in hand with the dedicated `v2.2.0` release of the AiiDA-VASP plugin, which is compatible
with `aiida_core<2`. This commit pins the AiiDA-VASP plugin to the `v2.2.0` version.