Dependencies: Restrict dependency version requirements
Packages that have a major release of v1 or higher are expected to
maintain semantic versioning and so packages are pinned to the major
version. Packages that haven't seen a v1 release yet, are pinned to the
minor version instead, as they are more likely to break backwards
compatibility in minor releases.