Dependencies: Make installing plugin packages optional
Installing all plugin packages is quite costly and most users are not
likely to want to use all codes. Therefore, an optional requirement is
created for each code that implements the common workflow interface.

To make it easy to still install all plugin packages, the `all_plugins`
optional requirement group adds the union of all the plugin optional
requirements. The downside is that this requires duplicating the
requirements and risks getting out of sync.

Built in `[all]` support was proposed in PEP 426 but this was rejected:
https://peps.python.org/pep-0426