Devops: Add pre-commit hook that validates optional dependencies
The `dev/validate_optional_dependencies.py` is added. It validates that
the `all_plugins` extras specifies exactly the same dependency
requirements that all other extras combined declare as well, except for
the `docs`, `pre-commit`, and `tests` extras, which are only used for
development.

This is to ensure that the `all_plugins` extras provides the exact same
dependencies as all the plugin specific extras combined. The script is
called through a pre-commit hook.