Devops: Add tests for the minimal install
A new job `tests-minimal-install` is added to the CI and CD workflows.
The job installs the package with minimal extra dependencies (just the
`tests` extras are installed so that `pytest` is available`. The job
then invokes `pytest` with the `-m minimal_install` option. The
`minimal_install` marker is added to run only those tests that should
run with a minimal install.

For now, the minimal install tests simply check that all modules are
importable, except for the workflow plugin implementations, and that all
CLI commands can be called with the `--help` option.