Devops: Exclude CP2K protocol YAML from format hook
The protocol file includes a couple of instances of `ON` and `OFF` that
are intended as literal string values. The formatter strips quotes from
strings without spaces as it is not required in YAML. However, in YAML,
the `ON` and `OFF` values are also aliases for `true` and `false` and so
the formatter standardizes these values to `true` and `false`. The file
is excluded from the formatter for that reason.