Fix: Set common inputs to `non_db` (#339)
Due to the updates in version 2.6 of `aiida-core`, the `to_aiida_type`
serializer is used as a default. Therefore, the validation fails for the
inputs defined in the common interfaces, as those are typically defined
as Python datatypes. Since these inputs are anyway not stored in the 
database, we fix it by explicitly setting `non_db = True` for those inputs.