Add support for fixed total magnetization and optional features
This PR adds support to run fixed spin moment calculations, i.e.,
enabling the user to specify the total magnetization of the cell that is
kepy during the calculation.

Moreover, since this is a new feature that is not supported by all the
plugins in the beginning (or maybe in general), this PR also introduces
the concept of additional optional features in the common workflows.
Once the plugin developers add support for that feature, it will be
enabled and is ready to use.