Add support for fixed total magnetization, optional features and small fixes (#344)
* Add support for fixed total magnetization and optional features

This PR adds support to run fixed spin moment calculations, i.e.,
enabling the user to specify the total magnetization of the cell that is
kept during the calculation.

Moreover, since this is a new feature that is not supported by all the
plugins in the beginning (or maybe in general), this PR also introduces
the concept of additional optional features in the common workflows.
Once the plugin developers add support for that feature, it will be
enabled and is ready to use.

* Fixes AiiDA-QE to be compatible with the latest changes there, in particular, the new protocol names.

* Small Read the Docs fix